Refactor(*): usecase 패키지의 상위에 application을 추가
