+++ jabiseo-api/src/main/java/com/jabiseo/learning/application/usecase/CreateLearningUseCase.java
@@ -1,4 +1,4 @@
-package com.jabiseo.learning.usecase;
+package com.jabiseo.learning.application.usecase;
 
 import com.jabiseo.learning.dto.CreateLearningRequest;
 import org.springframework.stereotype.Service;
