+++ jabiseo-api/src/main/java/com/jabiseo/member/application/usecase/FindMyInfoUseCase.java
@@ -1,4 +1,4 @@
-package com.jabiseo.member.usecase;
+package com.jabiseo.member.application.usecase;
 
 import com.jabiseo.member.domain.Member;
 import com.jabiseo.member.domain.MemberRepository;
