+++ jabiseo-api/src/main/java/com/jabiseo/member/controller/MemberController.java
@@ -5,9 +5,9 @@ import com.jabiseo.config.auth.AuthenticatedMember;
 import com.jabiseo.member.dto.FindMyCertificateStateResponse;
 import com.jabiseo.member.dto.FindMyInfoResponse;
 import com.jabiseo.member.dto.UpdateMyCertificateStateRequest;
-import com.jabiseo.member.usecase.FindMyCertificateStateUseCase;
-import com.jabiseo.member.usecase.FindMyInfoUseCase;
-import com.jabiseo.member.usecase.UpdateMyCertificateStateUseCase;
+import com.jabiseo.member.application.usecase.FindMyCertificateStateUseCase;
+import com.jabiseo.member.application.usecase.FindMyInfoUseCase;
+import com.jabiseo.member.application.usecase.UpdateMyCertificateStateUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
