+++ jabiseo-api/src/main/java/com/jabiseo/problem/controller/ProblemController.java
@@ -4,9 +4,9 @@ import com.jabiseo.problem.dto.CreateReportRequest;
 import com.jabiseo.problem.dto.FindBookmarkedProblemsResponse;
 import com.jabiseo.problem.dto.FindProblemsRequest;
 import com.jabiseo.problem.dto.FindProblemsResponse;
-import com.jabiseo.problem.usecase.CreateReportUseCase;
-import com.jabiseo.problem.usecase.FindBookmarkedProblemsUseCase;
-import com.jabiseo.problem.usecase.FindProblemsUseCase;
+import com.jabiseo.problem.application.usecase.CreateReportUseCase;
+import com.jabiseo.problem.application.usecase.FindBookmarkedProblemsUseCase;
+import com.jabiseo.problem.application.usecase.FindProblemsUseCase;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
