+++ jabiseo-api/src/test/java/com/jabiseo/certificate/application/usecase/FindCertificateListUseCaseTest.java
@@ -1,4 +1,4 @@
-package com.jabiseo.certificate.usecase;
+package com.jabiseo.certificate.application.usecase;
 
 import com.jabiseo.certificate.domain.Certificate;
 import com.jabiseo.certificate.domain.CertificateRepository;
