+++ jabiseo-api/src/test/java/com/jabiseo/problem/application/usecase/FindProblemsUseCaseTest.java
@@ -1,4 +1,4 @@
-package com.jabiseo.problem.usecase;
+package com.jabiseo.problem.application.usecase;
 
 import com.jabiseo.certificate.domain.Certificate;
 import com.jabiseo.certificate.domain.CertificateRepository;
